requirements.txt
boto==2.49.0
boto3==1.9.138
botocore==1.12.138
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-bootstrap-modal-forms==1.4.2
django-mathfilters==0.4.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.3
docutils==0.14
gensim==3.7.2
idna==2.8
jmespath==0.9.4
JPype1==0.6.3
konlpy==0.5.1
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
numpy==1.16.3
protobuf==3.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
smart-open==1.8.3
sqlparse==0.3.0
urllib3==1.24.2
